requirements.txt
awswrangler==3.9.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
torch==2.4.0
transformers==4.44.0
fastapi==0.112.0